Add Bovada Scrape
Add ==5 to bet type column check in case you get a wild game that is -1000s. Removed Consensus as it's not a bookie (lol). Added a line to bet type dataframes that remove incorrect line. Sometimes last row can be nulls.
(also renaming bovado_pull to bovada_pull)
Add webscrape for bovada. Scrape is in another file, referenced in main arb scanner. The bovada scrape pulls the right string 90% of the time. 10% of the time it pulls a string that doesnt have the info we need. Added a check for when this happens. If it happens, wait 2 seconds, then repeat. If it repeats for 60 seconds then skip the bovada pull
